Add stats helpers + documentation
